Adding .txt building to Makefile
Note: Making 'build-pdfs' the default instead of 'all' just because the
most common use-case will be checking that the PDF looks fine when
making changes to a resume. Since the .txt file requires manual editting
to make it look decent, that is the last step in making changes to the
resume.